Separated report tests into files
